remove NOJSLabelButton
